+++ GUI/coregui/Models/RectangularDetectorItem.cpp
@@ -146,11 +146,11 @@ std::unique_ptr<IDetector2D> RectangularDetectorItem::createDomainDetector() con
 {
     // basic axes parameters
     auto& x_axis = item<BasicAxisItem>(RectangularDetectorItem::P_X_AXIS);
-    int n_x = x_axis.getItemValue(BasicAxisItem::P_NBINS).toInt();
+    size_t n_x = x_axis.getItemValue(BasicAxisItem::P_NBINS).toUInt();
     double width = x_axis.getItemValue(BasicAxisItem::P_MAX).toDouble();
 
     auto& y_axis = item<BasicAxisItem>(RectangularDetectorItem::P_Y_AXIS);
-    int n_y = y_axis.getItemValue(BasicAxisItem::P_NBINS).toInt();
+    size_t n_y = y_axis.getItemValue(BasicAxisItem::P_NBINS).toUInt();
     double height = y_axis.getItemValue(BasicAxisItem::P_MAX).toDouble();
 
     std::unique_ptr<RectangularDetector> result(new RectangularDetector(n_x, width, n_y, height));
